invoice done

+++ SBW.Entities/SnPModule/InvoiceDetails.cs
@@ -12,9 +12,9 @@ namespace SBW.Entities.SnPModule
         public int invoiceNo { set; get; }
         public String productDescription { set; get; }
         public int qty { set; get; }
-        public float unitPrice { set; get; }
-        public float discount { set; get; }
+        public decimal unitPrice { set; get; }
+        public decimal discount { set; get; }
         public String warrenty { set; get; }
-        public float amount { set; get; }
+        public decimal amount { set; get; }
     }
 }
